Update license to GPLv3+
Readline was added almost 2 years ago in November 2013; because
readline is GPL-licensed and Sparkling links to libreadline, Sparkling
also must be GPL-licensed to comply with the license. Readline used to
be GPLv2+ licensed but is GPLv3+ licensed as of version 6.0, so use
GPLv3+ to to comply regardless of the version of readline.